requirements.txt
requests
pycryptodome
pyperclip
